+++ src/components/content/patients-adt/modal-hooks.tsx
@@ -90,8 +90,8 @@ export function useADTAlertDetailsDrawer() {
     subHeader: (adt) => `${adt.patient?.dob} (${adt.patient?.age})`,
     getSourceDocument: true,
     details: encounterData,
-    enableDismissAndReadActions: false, // We haven't gotten to supporting this yet
     renderChild: encountersAndNotes,
+    enableDismissAndReadActions: true,
   });
 }