+++ src/styles/tailwind.theme.ts
@@ -9,6 +9,7 @@ type Subset<K> = {
 export const CLASS_PREFIX = "ctw-";
 
 export const defaultBreakpoints = {
+  xs: 400,
   sm: 640,
   md: 768,
   lg: 1024,